Improve framework detection
For GOV.UK Frontend, additionally check for the `govuk-template__body` class on the `<body>` element.

For the legacy frameworks, additionally check for the header from GOV.UK Template.

Allow for the possibility that both GOV.UK Frontend and the legacy frameworks are in use at the same time.

Use `Array.prototype.some` to lazy evaluate the various checks.